CSP-610-PopulateProfileDataInSessionModel-CodeReview Fixes
