+++ src/SFA.DAS.ApprenticeAan.Web.UnitTests/Controllers/Onboarding/BeforeYouStartControllerTests/BeforeYouStartControllerPostTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
 using SFA.DAS.ApprenticeAan.Domain.Interfaces;
+using SFA.DAS.ApprenticeAan.Domain.OuterApi.Responses;
 using SFA.DAS.ApprenticeAan.Web.Controllers.Onboarding;
 using SFA.DAS.ApprenticeAan.Web.Infrastructure;
 using SFA.DAS.ApprenticeAan.Web.Models;
@@ -18,7 +19,7 @@ public class BeforeYouStartControllerPostTests
         [Frozen] Mock<ISessionService> sessionServiceMock,
         [Greedy] BeforeYouStartController sut)
     {
-        var result = await sut.PostAsync();
+        var result = await sut.Post();
 
         sessionServiceMock.Verify(s => s.Set(It.Is<OnboardingSessionModel>(m => !m.HasAcceptedTermsAndConditions)));
         result.As<RedirectToRouteResult>().RouteName.Should().Be("TermsAndConditions");
@@ -28,18 +29,18 @@ public class BeforeYouStartControllerPostTests
     public async Task Post_GetProfiles_UpdatesSessionModel(
         [Frozen] Mock<ISessionService> sessionServiceMock,
         [Frozen] Mock<IProfileService> profileServiceMock,
-        [Greedy] BeforeYouStartController sut)
+        [Greedy] BeforeYouStartController sut,
+        List<Profile> profiles)
     {
         const string userType = "apprentice";
 
         OnboardingSessionModel sessionModel = new();
-        var profiles = await profileServiceMock.Object.GetProfilesByUserType(userType);
 
         profileServiceMock.Setup(s => s.GetProfilesByUserType(userType)).ReturnsAsync(profiles);
         sessionServiceMock.Setup(s => s.Get<OnboardingSessionModel>()).Returns(sessionModel);
         sessionServiceMock.Object.Set(sessionModel);
 
-        await sut.PostAsync();
+        await sut.Post();
 
         sessionServiceMock.Verify(s => s.Set(sessionModel));
 
@@ -55,7 +56,7 @@ public class BeforeYouStartControllerPostTests
 
         sessionServiceMock.Setup(s => s.Get<OnboardingSessionModel>()).Returns(sessionModel);
 
-        var result = await sut.PostAsync();
+        var result = await sut.Post();
 
         result.As<RedirectToRouteResult>().Should().NotBeNull();
         result.As<RedirectToRouteResult>().RouteName.Should().Be(RouteNames.Onboarding.TermsAndConditions);
