+++ src/SFA.DAS.ApprenticeAan.Web/Models/OnboardingSessionModel.cs
@@ -1,6 +1,4 @@
-﻿using SFA.DAS.ApprenticeAan.Domain.OuterApi.Responses;
-
-namespace SFA.DAS.ApprenticeAan.Web.Models;
+﻿namespace SFA.DAS.ApprenticeAan.Web.Models;
 
 public class OnboardingSessionModel
 {
@@ -19,21 +17,4 @@ public class ApprenticeDetailsModel
     public string? Email { get; set; }
     public string? Name { get; set; }
     public string? ReasonForJoiningTheNetwork { get; set; }  //maps to information field in apprentice table
-}
-
-public class ProfileModel
-{
-    public int Id { get; set; }
-    public string Value { get; set; } = null!;
-    public string Description { get; set; } = null!;
-    public string Category { get; set; } = null!;
-    public int Ordering { get; set; }
-
-    public static implicit operator ProfileModel(Profile profile) => new()
-    {
-        Id = profile.Id,
-        Description = profile.Description,
-        Category = profile.Category,
-        Ordering = profile.Ordering
-    };
 }
