+++ src/SFA.DAS.ApprenticeAan.Web/Models/ProfileModel.cs
@@ -1,0 +1,20 @@
+﻿using SFA.DAS.ApprenticeAan.Domain.OuterApi.Responses;
+
+namespace SFA.DAS.ApprenticeAan.Web.Models;
+
+public class ProfileModel
+{
+    public int Id { get; set; }
+    public string Value { get; set; } = null!;
+    public string Description { get; set; } = null!;
+    public string Category { get; set; } = null!;
+    public int Ordering { get; set; }
+
+    public static implicit operator ProfileModel(Profile profile) => new()
+    {
+        Id = profile.Id,
+        Description = profile.Description,
+        Category = profile.Category,
+        Ordering = profile.Ordering
+    };
+}
